Communication binaire débuguée Commandes com binaire renomée: ask: tout ce qui renvoie des données set: tout ce qui modifice des valeurs le reste: des actions
Ajout de commande:
Action sur pompe
Retourne les 8 adc
ask position (qui n'est plus automatique)

+++ modules/cvra_adc/cvra_adc.c
@@ -71,7 +71,7 @@ void cvra_adc_start_scan(cvra_adc_t *adc) {
 	IOWR(adc->spi_adress, TXDATA_REGISTER, (adc->next_read)<<11);
 }
 
-int cvra_adc_get_value(cvra_adc_t *adc, int input) {
+int16_t cvra_adc_get_value(cvra_adc_t *adc, int input) {
     if(input >= 8 || input < 0)
         return -1;
     return adc->values[input];
